pyproject.toml
[project]
name = "clickspeak"
version = "0.1.0"
description = "Keyboard-free voice-to-text for macOS"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pyobjc-core",
    "pyobjc-framework-Cocoa",
    "pyobjc-framework-ApplicationServices",
    "pyobjc-framework-Quartz",
    "parakeet-mlx",
    "sounddevice",
    "openwakeword",
    "onnxruntime",
    "pynput",
    "rumps",
    "soundfile",
    "numpy",
    "numba>=0.60",
    "llvmlite>=0.43",
]

[project.scripts]
clickspeak = "clickspeak.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clickspeak"]
